refactor: 💡 search input
