+++ src/components/SearchBar/__tests__/SearchBar.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import SearchBar from '..';
+import SearchInput from '..';
 
 describe('Searbar component', () => {
-  it.only('should render correctly', () => {
-    const component = render(<SearchBar />);
+  it('should render correctly', () => {
+    const component = render(<SearchInput />);
     expect(component).toMatchSnapshot();
   });
 });
