+++ src/components/SearchBar/index.tsx
@@ -1,20 +1,17 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useFlexSearch } from 'react-use-flexsearch';
-import { AnimatePresence, motion } from 'framer-motion';
-import { useClickOutside } from 'react-click-outside-hook';
-import { graphql, Link, useStaticQuery } from 'gatsby';
-import { SearchResult } from '../../types';
-import './SearchBar.scss';
+import { graphql, useStaticQuery } from 'gatsby';
 
-const containerTransition = { type: 'spring', damping: 22, stiffness: 150 };
-const containerVariants = {
-  expanded: {
-    height: '30em',
-  },
-  collapsed: {
-    height: '3.8em',
-  },
-};
+import './SearchBar.scss';
+import SearchInput from './SearchInput';
+
+export interface SearchProps {
+  localSearchLearnPages: {
+    id: string;
+    store: string;
+    index: string[];
+  };
+}
 const SearchBar = (): JSX.Element => {
   const queryData = useStaticQuery(graphql`
     query {
@@ -29,100 +26,21 @@ const SearchBar = (): JSX.Element => {
   const { index } = queryData.localSearchLearnPages;
   const { store } = queryData.localSearchLearnPages;
   const [query, setQuery] = useState('');
-  const [isExpanded, setExpanded] = useState(false);
-  const [parentRef, isClickedOutside] = useClickOutside();
   const [noResults, setNoResults] = useState(false);
 
   const results = useFlexSearch(query, index, store);
 
   const isEmpty = !results || results.length === 0;
 
-  const changeHandler = (e: {
-    preventDefault: () => void;
-    target: { value: React.SetStateAction<string> };
-  }) => {
-    e.preventDefault();
-    if (e.target.value === '') {
-      setNoResults(false);
-    }
-    setQuery(e.target.value);
-  };
-
-  const expandContainer = () => {
-    setExpanded(true);
-  };
-
-  const collapseContainer = () => {
-    setExpanded(false);
-    setQuery('');
-    setNoResults(false);
-  };
-
-  useEffect(() => {
-    if (isClickedOutside) {
-      collapseContainer();
-    }
-  }, [isClickedOutside]);
-
   return (
-    <motion.div
-      className="searchBarContainer"
-      animate={isExpanded ? 'expanded' : 'collapsed'}
-      variants={containerVariants}
-      transition={containerTransition}
-      ref={parentRef}
-    >
-      <div className="searchInputContainer">
-        <div className="searchIcon material-icons">search</div>
-        <input
-          autoComplete="off"
-          className="inputText"
-          name="query"
-          value={query}
-          onChange={changeHandler}
-          placeholder="search"
-          onFocus={expandContainer}
-        />
-        <AnimatePresence>
-          {isExpanded && (
-            <motion.span
-              className="material-icons closeIcon"
-              key="close-icon"
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-              onClick={collapseContainer}
-              transition={{ duration: 0.2 }}
-            >
-              close
-            </motion.span>
-          )}
-        </AnimatePresence>
-      </div>
-
-      {isExpanded && (
-        <div className="searchContent">
-          {/* loading state with an expanded area for the results */}
-          {isEmpty && !noResults && (
-            <div className="loadingWrapper">
-              <div className="warningMessage">Start typing to Search</div>
-            </div>
-          )}
-          {/* success state */}
-          {!isEmpty && (
-            <>
-              {results.map((result: SearchResult) => (
-                <ul key={result.id}>
-                  <li className="resultItem">
-                    <Link to={result.slug}>{result.title}</Link>
-                  </li>
-                </ul>
-              ))}
-            </>
-          )}
-        </div>
-      )}
-    </motion.div>
+    <SearchInput
+      setQuery={setQuery}
+      isEmpty={isEmpty}
+      noResults={noResults}
+      setNoResults={setNoResults}
+      query={query}
+      results={results}
+    />
   );
 };
 export default SearchBar;
